feat: Add EPA field for projects and activities
Introduce a new boolean field `isEPA` to the project and activity types. This field will be used to distinguish projects or activities that are part of an "Environmental, Social, and Governance" (ESG) initiative.

The `ProjectsExplorer` and `Projects` pages have been updated to display and allow filtering/editing of this new field. The `ProjectsExplorer` type has also been updated to include `isEPA`. Sample data in `data/projects.ts` has been updated to include `isEPA: false` for existing projects.

Additionally, a new search term for "orientation" has been added to the `Activities` page to support future filtering based on project orientations.